Consultas y DML
Creación de un archivo DML para ingresar algunos datos de prueba y un archivo con las consultas para probar la base.

+++ Proyecto Integrador/proyecto/src/main/java/ar/org/centro8/curso/java/proyecto/sql/base.sql
@@ -12,10 +12,10 @@ create table productos(
 	id int auto_increment primary key,
     nombre varchar(100) not null,
     marca varchar(40) not null,
-    tipoProducto enum("CARNES", "VERDURAS", "PANIFICADOS", "FRUTAS", "LACTEOS", "BEBIDAS"),
+    tipoProducto enum('CARNES', 'VERDURAS', 'PANIFICADOS', 'FRUTAS', 'LACTEOS', 'BEBIDAS'),
     precioCosto int not null,
-    unidad enum("KG", "LT", "UNIDAD"),
-    cantidadPorUnidad int not null
+    unidad enum('KG', 'LT', 'UNIDAD'),
+    cantidadPorUnidad numeric(6,2) not null
 );
 
 create table colegios(
